feat: RAM editor (iOS), Java version-picker installer, iOS tab reorg + runtime switching
Add /config/ram GET/POST to the Remote API and a Memory editor card in the
iOS Health tab. Java servers expose Min/Max (-Xms/-Xmx) steppers; Bedrock
gets a single VM memory limit. Values are clamped server-side
(0.5 ≤ min ≤ max ≤ physical RAM) and persisted to the active server config;
admin/settings-permission only, guests read-only.

Add an Install Java flow to Settings and the Setup Wizard that lets users
pick a Java version framed by Minecraft compatibility, then downloads and
auto-opens the Adoptium Temurin installer to ~/Downloads. Corrects the
Minecraft-to-Java mapping now that MC 26.1 requires Java 25.

Reorganise iOS tabs, expand the server creation flow, add Java runtime
switching, and add Xbox broadcast JAR management.

+++ MSCmacOS/MSCmacOS Swift/DetailsPerformanceComponents.swift
@@ -34,11 +34,11 @@ extension DetailsPerformanceTabView {
         return String(format: "%.0f%%", v)
     }
 
-    func formatRam(_ mb: Double?, maxGB: Int?) -> String {
+    func formatRam(_ mb: Double?, maxGB: Double?) -> String {
         guard let mb = mb else { return "--" }
 
         if let maxGB = maxGB {
-            let maxMB = Double(maxGB) * 1024.0
+            let maxMB = maxGB * 1024.0
             return String(format: "%.1f / %.0f MB", mb, maxMB)
         } else {
             return String(format: "%.1f MB", mb)
